requirements.txt
PyQt6
opencv-python
pyzbar
Pillow
numpy
